Phase 4a: extract NervesHub.Devices.CACertificates sub-context
First step of splitting the Devices god-module (129 public fns, ~1900 lines)
into focused sub-contexts. Built on Phase 3 (cleanup/phase-3-encapsulation).

Moves the 10 CA-certificate functions (create/get/update/delete_ca_certificate,
create_ca_certificate_from_x509, the by-aki/ski/serial/org_and_serial getters,
and known_ca_ski?) out of NervesHub.Devices into a new, documented
NervesHub.Devices.CACertificates module. These operate on Org/CACertificate and
have no Device involvement, so they were the most self-contained sub-domain.

Function names are unchanged; only the module qualifier moved. All 46 call sites
across lib and test were repointed to the new module, dropping now-unused
`Devices` aliases where applicable. `join`/query helpers stay internal.

Verification: mix format, credo (new module clean), dialyzer (0 unskipped),
full test suite (1396 passing).
